extract MathPreview
